fix(vue_archi): refactorisation
