+++ yaki_admin/src/features/invitation/components/UserInvitationCard.vue
