+++ yaki_admin/src/models/teammate.type.ts
@@ -1,0 +1,10 @@
+export type TeammateTypeIn = {
+  id: number;
+  teamId: number;
+  userId: number;
+};
+
+export type TeammateTypeOut = {
+  teamId: number;
+  userId: number;
+};
